// meal-diary-frontend/src/graphql/generated.tsx
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type AdditionalEntityFields = {
  path?: Maybe<Scalars['String']>;
  type?: Maybe<Scalars['String']>;
};


export type Config = {
  __typename?: 'Config';
  lastFineliUpdate?: Maybe<Scalars['Date']>;
};

export type SearchFloatInput = {
  value: Scalars['Float'];
  comparisonOperator?: Maybe<ComparisonOperator>;
};

export enum ComparisonOperator {
  Lt = 'LT',
  Gt = 'GT'
}

export type SearchStringInput = {
  value: Scalars['String'];
  regex?: Maybe<Scalars['Boolean']>;
  flags?: Maybe<Scalars['String']>;
};

export type User = {
  __typename?: 'User';
  name?: Maybe<Scalars['String']>;
  userId: Scalars['String'];
  roles?: Maybe<Array<Role>>;
  permissions?: Maybe<Array<Permission>>;
};
// ...
export type Query = {
  __typename?: 'Query';
  users?: Maybe<Array<User>>;
  foods: Array<Food>;
  food?: Maybe<Food>;
  config: Config;
};
// ...
export type FoodName = {
  __typename?: 'FoodName';
  en?: Maybe<Scalars['String']>;
  fi?: Maybe<Scalars['String']>;
};

export type Food = {
  __typename?: 'Food';
  id: Scalars['ID'];
  name: FoodName;
  enerc?: Maybe<Scalars['Float']>;
  fat?: Maybe<Scalars['Float']>;
  choavl?: Maybe<Scalars['Float']>;
  prot?: Maybe<Scalars['Float']>;
  alc?: Maybe<Scalars['Float']>;
  oa?: Maybe<Scalars['Float']>;
  sugoh?: Maybe<Scalars['Float']>;
  sugar?: Maybe<Scalars['Float']>;
  frus?: Maybe<Scalars['Float']>;
  gals?: Maybe<Scalars['Float']>;
  glus?: Maybe<Scalars['Float']>;
  lacs?: Maybe<Scalars['Float']>;
  mals?: Maybe<Scalars['Float']>;
  sucs?: Maybe<Scalars['Float']>;
  starch?: Maybe<Scalars['Float']>;
  fibc?: Maybe<Scalars['Float']>;
  fibins?: Maybe<Scalars['Float']>;
  psacncs?: Maybe<Scalars['Float']>;
  fol?: Maybe<Scalars['Float']>;
  niaeq?: Maybe<Scalars['Float']>;
  nia?: Maybe<Scalars['Float']>;
  vitpyrid?: Maybe<Scalars['Float']>;
  ribf?: Maybe<Scalars['Float']>;
  thia?: Maybe<Scalars['Float']>;
  vita?: Maybe<Scalars['Float']>;
  carotens?: Maybe<Scalars['Float']>;
  vitb12?: Maybe<Scalars['Float']>;
  vitc?: Maybe<Scalars['Float']>;
  vitd?: Maybe<Scalars['Float']>;
  vite?: Maybe<Scalars['Float']>;
  vitk?: Maybe<Scalars['Float']>;
  ca?: Maybe<Scalars['Float']>;
  fe?: Maybe<Scalars['Float']>;
  iod?: Maybe<Scalars['Float']>;
  k?: Maybe<Scalars['Float']>;
  mg?: Maybe<Scalars['Float']>;
  na?: Maybe<Scalars['Float']>;
  nacl?: Maybe<Scalars['Float']>;
  p?: Maybe<Scalars['Float']>;
  se?: Maybe<Scalars['Float']>;
  zn?: Maybe<Scalars['Float']>;
  fafre?: Maybe<Scalars['Float']>;
  fapu?: Maybe<Scalars['Float']>;
  famcis?: Maybe<Scalars['Float']>;
  fasat?: Maybe<Scalars['Float']>;
  fatrn?: Maybe<Scalars['Float']>;
  fapun3?: Maybe<Scalars['Float']>;
  fapun6?: Maybe<Scalars['Float']>;
  f18d2cn6?: Maybe<Scalars['Float']>;
  f18d3n3?: Maybe<Scalars['Float']>;
  f20d5n3?: Maybe<Scalars['Float']>;
  f22d6n3?: Maybe<Scalars['Float']>;
  chole?: Maybe<Scalars['Float']>;
  stert?: Maybe<Scalars['Float']>;
  trp?: Maybe<Scalars['Float']>;
  foodid?: Maybe<Scalars['String']>;
};
// ...
export type GetFoodQuery = (
  { __typename?: 'Query' }
  & { food?: Maybe<(
    { __typename?: 'Food' }
    & Pick<Food, 'id' | 'enerc' | 'fat' | 'prot' | 'sugar' | 'fibc' | 'nacl'>
    & { name: (
      { __typename?: 'FoodName' }
      & Pick<FoodName, 'fi' | 'en'>
    ) }
  )> }
);
// ...
export const GetFoodDocument = gql`
    query getFood($id: ID!) {
  food(id: $id) {
    id
    name {
      fi
      en
    }
    enerc
    fat
    prot
    sugar
    fibc
    nacl
  }
}
    `;